move logic to correct location
